Task: added blueprint to organize routes and make apis accessible from routes.py in app.py
